fix(*): restructure components
