+++ dev/serve-components/index.ts
@@ -8,7 +8,7 @@ const app = createApp(ComponentsApp)
 const pluginOptions: KongAuthElementsOptions = {
   // Unless using an absolute URL, this base path MUST start with a leading slash (if setting the default) in order to properly resolve within container applications, especially when called from nested routes(e.g. /organizations/users)
   apiBaseUrl: '/kauth',
-  userEntity: 'developer',
+  userEntity: 'user',
   customErrorHandler: ({ error, request, element }): string => {
     console.log('error', error)
