+++ src/assets/styles/app.scss
@@ -25,8 +25,7 @@
   }
 }
 
-.kong-auth-element,
-.base-custom-element {
+.kong-auth-element {
   // Import utility classes from Kongponents, but namespace them so they don't leak into page
   @import "@kongponents/styles/utils";
