+++ src/components/BaseCustomElement.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="base-custom-element">
+  <div class="kong-auth-element">
     <div data-testid="kong-auth-injected-styles" v-if="injectedStyles">
       <!-- This injected styles tag, the parent tag, and its corresponding logic in the setup function must be present to import styles from child components. -->
       <div v-html="injectedStyles"></div>
