+++ src/elements/kong-auth-forgot-password/KongAuthForgotPassword.spec.ts
@@ -22,17 +22,13 @@ const user = {
   email: 'user1@email.com',
 }
 
-describe('KongAuthForgotPassword.vue', () => {
+describe('KongAuthForgotPassword.ce.vue', () => {
   // Required for all Custom Elements
   it('has proper structure and required classes', () => {
     mount(KongAuthForgotPassword)
 
     // Should have .kong-auth-element as parent class
-    const elementDiv = cy.get('.kong-auth-element')
-    elementDiv.should('not.be.empty')
-
-    // Should have BaseCustomElement as a child component
-    elementDiv.find('.base-custom-element').should('be.visible').and('not.be.empty')
+    cy.get('.kong-auth-element').should('be.visible').and('not.be.empty')
   })
 
   it('renders a forgot password form with email and button elements', () => {
